requirements.txt
notion-client>=2.0.0
python-dotenv>=1.0.0
yt-dlp>=2023.3.4
ffmpeg-python>=0.2.0
requests>=2.28.2
pathlib>=1.0.1
typing-extensions>=4.5.0
instaloader>=4.10.3 